+++ Shuhari.Framework.Testing.Common/Properties/AssemblyInfo.cs
@@ -2,7 +2,7 @@
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("Shuhari.Framework.Testing.NUnit")]
+[assembly: AssemblyTitle("Shuhari.Framework.Testing.Common")]
 [assembly: AssemblyDescription("Testing helpers for NUnit")]
 [assembly: Guid("30b90498-913b-41a9-9022-e5df25c59e37")]