+++ Shuhari.Framework.Testing.Mvc/Properties/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+﻿using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("Shuhari.Framework.Testing.Mvc")]
+[assembly: AssemblyDescription("Testing helpers for ASP.NET MVC")]
+[assembly: Guid("abbd3c1f-977b-460a-8b9d-9d6fe0d39eda")]