+++ Shuhari.Framework.Wpf/Controls/HBox.cs
@@ -1,0 +1,18 @@
+﻿using System.Windows.Controls;
+
+namespace Shuhari.Framework.Wpf.Controls
+{
+    /// <summary>
+    /// Horizontal box
+    /// </summary>
+    public class HBox : Box
+    {
+        /// <summary>
+        /// Initialize
+        /// </summary>
+        public HBox()
+        {
+            Orientation = Orientation.Horizontal;
+        }
+    }
+}