+++ Shuhari.Framework.Wpf/Controls/VBox.cs
@@ -1,0 +1,18 @@
+﻿using System.Windows.Controls;
+
+namespace Shuhari.Framework.Wpf.Controls
+{
+    /// <summary>
+    /// Vertial box
+    /// </summary>
+    public class VBox : Box
+    {
+        /// <summary>
+        /// Initialize
+        /// </summary>
+        public VBox()
+        {
+            Orientation = Orientation.Vertical;
+        }
+    }
+}