+++ Shuhari.Framework.Wpf/Properties/AssemblyInfo.cs
@@ -1,6 +1,30 @@
 ﻿using System.Reflection;
 using System.Runtime.InteropServices;
+using System.Windows.Markup;
+using Shuhari.Framework.Wpf;
 
 [assembly: AssemblyTitle("Shuhari.Framework.Wpf")]
 [assembly: AssemblyDescription("WPF Extensions")]
 [assembly: Guid("f42283da-f5dc-4f3e-89b8-e5d0e25043c6")]
+
+[assembly: XmlnsDefinition(AssemblyInfo.XAML_NAMESPACE, "Shuhari.Framework.Wpf.Controls")]
+[assembly: XmlnsPrefix(AssemblyInfo.XAML_NAMESPACE, AssemblyInfo.XAML_PREFIX)]
+
+namespace Shuhari.Framework.Wpf
+{
+    /// <summary>
+    /// Assembly constants
+    /// </summary>
+    public static class AssemblyInfo
+    {
+        /// <summary>
+        /// XAML namespace
+        /// </summary>
+        public const string XAML_NAMESPACE = "https://shuhari.github.io/xaml";
+
+        /// <summary>
+        /// XAML prefix
+        /// </summary>
+        public const string XAML_PREFIX = "sw";
+    }
+}